+++ src/main/java/app/IApp.java
@@ -3,7 +3,7 @@ package app;
 import data_access.IDataAccessor;
 import entity.calculator.IHandStateFactory;
 import entity.user.IUserManager;
-import use_case.ViewManager;
+import interface_adapter.ViewManager;
 import view.AbstractPanel;
 
 public interface IApp {
